:art: onBtnCarNamesClicked() 모듈 이동
