+++ src/library/car-names.js
@@ -1,6 +1,3 @@
-import { makeElementBlock } from "./make-element-block.js";
-import { Car } from "./car.js";
-
 export function isCorrectCarNames(carName) {
   if (carName.length > 0 && carName.length <= 5) {
     return true;
@@ -23,19 +20,4 @@ export function addCorrectCarNames(carNames, inputCarNames) {
   });
 
   return correctCarNames;
-}
-
-export function onBtnCarNamesClicked(cars, inputCarNames) {
-  const carNames = inputCarNames.value.split(",");
-  const correctCarNames = addCorrectCarNames(carNames, inputCarNames);
-  const racingBox = document.getElementById("racing-count-box");
-
-  if (!correctCarNames.includes(undefined)) {
-    makeElementBlock(racingBox);
-    correctCarNames.forEach((correctCar) => {
-      cars.push(new Car(correctCar));
-    });
-  }
-
-  return;
 }
